+++ Src/Files/FRandomLoader.hpp
@@ -34,7 +34,7 @@
 template <class ParticleClass>
 class FRandomLoader : public FAbstractLoader<ParticleClass> {
 protected:
-    const int nbParticles;            //< the number of particles
+    const size_t nbParticles;            //< the number of particles
     const FReal boxWidth;             //< the box width
     const FPoint centerOfBox;    //< The center of box
 
@@ -42,7 +42,7 @@ public:
     /**
     * The constructor need the simulation data
     */
-    FRandomLoader(const int inNbParticles, const FReal inBoxWidth = 1.0,
+    FRandomLoader(const size_t inNbParticles, const FReal inBoxWidth = 1.0,
                   const FPoint& inCenterOfBox = FPoint(0,0,0), const unsigned int inSeed = static_cast<unsigned int>(0))
         : nbParticles(inNbParticles), boxWidth(inBoxWidth), centerOfBox(inCenterOfBox) {
         srand(inSeed);
@@ -110,7 +110,7 @@ public:
 template <class ParticleClass>
 class FRandomLoaderTsm : public FRandomLoader<ParticleClass> {
 public:
-    FRandomLoaderTsm(const int inNbParticles, const FReal inBoxWidth = 1.0,
+    FRandomLoaderTsm(const size_t inNbParticles, const FReal inBoxWidth = 1.0,
                   const FPoint& inCenterOfBox = FPoint(0,0,0), const unsigned int inSeed = static_cast<unsigned int>(time(NULL)))
         : FRandomLoader<ParticleClass>(inNbParticles,inBoxWidth,inCenterOfBox,inSeed) {
     }
